Rename parser.py to avoid name conflicts, create directories if they don't exist yet
